requirements.txt
Flask==3.0.3
flask_cors==5.0.0
opencv-python
deepface
tf_keras
python-dotenv